requirements.txt
python-dotenv==1.0.0
tweepy==4.14.0 